Separating matches, and making them ConsoleMatches
